requirements.txt
asgiref==3.6.0
Django==4.1.7
Pillow==9.4.0
psycopg2==2.9.5
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
